Extract civilization editor.
